minimalistcss
This Sample Webpage Is An Example of Using Minimal CSS to Format a Webpage.
This webpage design was inspired by a video by Kevin Powell titled "Stop over-engineering your CSS". This entire webpage is built using less than 40 lines of CSS.
This webpage also features the use of two cool concepts: the min() Function, and the color-scheme property.